refactor: split notices into TravelCardNotices.tsx
